Prevent dependency issue
